backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
pydantic>=2.11.5
llama-index-core>=0.14.0
llama-index-llms-gemini>=0.6.0
llama-index-llms-openai>=0.6.0
llama-index-embeddings-openai>=0.3.0
llama-index-readers-file>=0.4.0
xhtml2pdf==0.2.16
fpdf2==2.8.3
python-dotenv==1.1.0
supabase>=2.10.0
openai>=1.0.0
python-telegram-bot==21.0